Refactored structure
